umkm-pos-ml/apps/requirements.txt
django==5.1.*
djangorestframework==3.15.*
django-cors-headers==4.4.*
psycopg2-binary==2.9.*
python-dotenv==1.0.*
requests==2.32.*

# Machine Learning
scikit-learn==1.5.*
pandas==2.2.*
numpy==2.0.*

# PDF Generation
reportlab==4.2.*

# Task scheduling
django-crontab==0.7.*

# Storage
boto3==1.35.*